Change what happens to a ship in the gameboard when hit
Instead of changing the array to "X" for a ship in a given index of the gameboard array, keep the array
in order to maintain information about the ship in the gameboard. Simply change the ship  index reference in position
[1] to be an "X", which means that part of the ship has been hit on the gameboard. This will help with styling
and relaying ship information to the DOM even after a part of a ship has been hit or sunk.